Ajout d'une fenêtre de dialogue et de sa classe associée

+++ sources/SQL_requests.py
@@ -113,16 +113,17 @@ class Database:
         self.c=self.database.cursor()
         self.c.execute("""SELECT COUNT(id_pokemon) FROM Equipe WHERE id_joueur=?""",(id_joueur,))
         self.results=self.c.fetchall()
-        print(self.results)
         if self.results[0][0]<6:
-            self.c.execute("""SELECT id_pokemon FROM Equipe WHERE id_joeuur=?""",(id_joueur,))
+            self.c.execute("""SELECT id_pokemon FROM Equipe WHERE id_joueur=?""",(id_joueur,))
             self.results=self.c.fetchall()
-            print(self.results)
-            self.info_pokemon=self.get_info_pokemon(id_pokemon)
-            self.c.execute("""INSERT INTO Equipe VALUES (?,?)""",(id_pokemon,self.info_pokemon["ID"]))
-            self.database.commit()
-            self.c.execute("""DELETE FROM PC WHERE id_pokemon=?""",(id_pokemon,))
-            self.database.commit()
+            if id_pokemon not in self.results[0]:
+                self.info_pokemon=self.get_info_pokemon(id_pokemon)
+                self.c.execute("""INSERT INTO Equipe VALUES (?,?)""",(id_pokemon,self.info_pokemon["ID"]))
+                self.database.commit()
+                self.c.execute("""DELETE FROM PC WHERE id_pokemon=?""",(id_pokemon,))
+                self.database.commit()
+            else:
+                return None
         else :
             return None
